services/inventory-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Marshmallow==1.0.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
